+++ chandler/repository/tests/RepositoryTestCase.py
@@ -6,18 +6,20 @@ __date__      = "$Date$"
 __copyright__ = "Copyright (c) 2003 Open Source Applications Foundation"
 __license__   = "http://osafoundation.org/Chandler_0.1_license_terms.htm"
 
-import unittest, os, sys
+from unittest import TestCase
+import os, sys
 
 from repository.persistence.XMLRepository import XMLRepository
 
-class RepositoryTestCase(unittest.TestCase):
+class RepositoryTestCase(TestCase):
 
-    def setUp(self):
+    def setUp(self, ramdb=False):
+        self.ramdb = ramdb
         self.rootdir = os.environ['CHANDLERHOME']
         self.testdir = os.path.join(self.rootdir, 'chandler', 'repository',
                                     'tests')
         self.rep = XMLRepository(os.path.join(self.testdir, '__repository__'))
-        self.rep.create()
+        self.rep.create(ramdb=self.ramdb)
         schemaPack = os.path.join(self.rootdir, 'chandler', 'repository',
                                   'packs', 'schema.pack')
         self.rep.loadPack(schemaPack)
@@ -25,13 +27,20 @@ class RepositoryTestCase(unittest.TestCase):
 
     def tearDown(self):
         self.rep.close()
-        self.rep.delete()
+        if not self.ramdb:
+            self.rep.delete()
 
     def _reopenRepository(self):
         self.rep.commit()
-        self.rep.close()
-        self.rep = XMLRepository(os.path.join(self.testdir, '__repository__'))
-        self.rep.open()
+
+        if self.ramdb:
+            self.rep.closeView()
+            self.rep.openView()
+        else:
+            self.rep.close()
+            self.rep = XMLRepository(os.path.join(self.testdir,
+                                                  '__repository__'))
+            self.rep.open()
 
     def _find(self, item):
         return self.rep.find(item)
